requirements.txt
pandas
tqdm
requests
pillow
transformers
